+++ src/BlockchainCommon/Common/VectorOutputStream.cs
@@ -8,20 +8,20 @@ namespace Common
 
 public class VectorOutputStream : IOutputStream
 {
-  public VectorOutputStream(List<uint8_t> @out)
+  public VectorOutputStream(List<ushort> @out)
   {
 	  this.@out = @out;
   }
 //C++ TO C# CONVERTER TODO TASK: C# has no equivalent to ' = delete':
 //  VectorOutputStream& operator =(const VectorOutputStream&) = delete;
-  public override uint64_t writeSome(object data, uint64_t size)
+  public override ulong writeSome(object data, ulong size)
   {
 //C++ TO C# CONVERTER TODO TASK: There is no direct equivalent to the STL vector 'insert' method in C#:
-	@out.insert(@out.end(), (uint8_t)data, (uint8_t)data + size);
+	@out.insert(@out.end(), (ushort)data, (ushort)data + size);
 	return size;
   }
 
-  private List<uint8_t> @out;
+  private List<ushort> @out;
 }
 
 }
